fixed wrong int

+++ pkg/raft/raftstate/raftstate.go
@@ -19,9 +19,9 @@ func (r *RaftState) SetState(s State) {
 	atomic.StoreUint32(stateAddr, uint32(s))
 }
 
-func (r *RaftState) GetTerm() int {
+func (r *RaftState) GetTerm() uint64 {
 	termAddr := (*uint64)(&r.currTerm)
-	return int(atomic.LoadUint64(termAddr))
+	return atomic.LoadUint64(termAddr)
 }
 
 func (r *RaftState) SetTerm(newTerm uint64) {
